ath12k: wifi: Correct mcs_nss_support_bw for EHT mode
In order for calculations for nss correct,
mcs_nss_support_bw needs to point to the correct table.

mcs_nss_support_bw should default back to 80mhz,
which is not the case.

There is a situation in the case statement combined
with the fallthrough which will allow the default
to be 160 mhz.

This can cause nss calculations of 0 where 160mhz
may not be available.

This change is to fix a regression, since this code
is missing from package upgrade.

Change-Id: I2cf1c930e1407801676ca175ad1a73750aa98746
Signed-off-by: Thomas Wu <quic_wthomas@quicinc.com>