wifi: ath12k: Send wdev in AFC power update complete event
In a scenario within a wiphy where there are two AP MLDs with two links
operating on 2 GHz and 5 GHz, and one 6 GHz SLO interface, the AP is
unable to come on SP mode after AFC is complete on 6 GHz due to a
failure in the set_6ghz_power_mode command.

An error message print: “Standard Power mode can’t be set without AFC.”
is observed. This occurs because the AFC power update event is sent to
the hostapd without wdev as a parameter. As a result, the AFC response
data is sent to the incorrect interface in the hostapd.

To fix this issue, in ath12k_vendor_send_power_update_complete
call cfg80211_vendor_event_alloc with wdev as a parameter.

Tag: 6G_REGULATORY
Fixes: Ic35d5d81c7b ("wifi: ath12k: Extract AFC response and send AFC cmd to FW")

Change-Id: Ida667cfa9e774cf9197f8ad387aa40dffc413655
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>