QSDK: ath12k: Fix invalid usage of ahvif::primary_link_id
When reusing vp profiles for split phy, primary goal is to get primary
link arvif from ahvif for the STA mode. For this ahvif::primary_link_id has been used which caused the NULL pointer dereference.

ahvif::primary_link_id is a user configured value and will have default value (0) if not configured. But it is not necessary that link 0 in ahvif has to be valid in all cases.

Conside below case:
1. Station assiciated in 2 link MLD 5GHz/6GHz with links map 0x6.
2. Hence valid links are 1,2
3. If primary link_id is used to get arvif, this will cause NULL pointer
deference.

Fix this by using ahsta::primary_link_id in case of MLO.

Fixes: If9c211d6097 (QSDK: ath12k: Reuse vp profiles in case of
splitphy)

Patch-dependency: 970-QSDK-ath12k-Reuse-vp-profiles-in-case-of-splitphy.patch

Change-Id: I1f1c2c45d028c27bed59e764490a3e6a552d258f
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>