QSDK: wifi: ath12k: Set wdev VP type at start in vendor command
Currently wdev VP type is set at the end of vendor command handler API.
Due to the timing issue, by the time vendor command is received,
sdata might not be running and it will return without updating the
VP type.

Hence setting the wdev VP early in the vendor command handler so that
at alloc, correct wdev type will be fetched and futher VP update
is not needed.

Change-Id: I66a4d659572d56862572a1ea74c47656b52cc845
Signed-off-by: Parikshit <quic_pgune@quicinc.com>