wifi: ath12k: split the scan req for 6GHz low and 6GHz high bands
When supplicant initiates scan req for low and high band device,
mac80211 generates an single scan req to driver with all the
channel frequencies. This is because mac80211 categorize the
the frequencies based on band only.

Driver checks the first freq from the channel list and initiates
scan to firmware with all the frequencies. Firmware rejects the freq
which are not supported. As a result of this, Supplicant receives
the scan results from only one device and associates on one link.

Fix this by splitting the scan request in driver and perform
multiple scans, send completion notification to mac80211 after
split scans are done.

Patch-work: None
Patch-dependency: 652-04-ath12k-Support-scan-for-single-wiphy-architecture.patch

Change-Id: I73cdf5592b2a66986f28eae056254937877f5101
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>