wifi: ath12k: Drop mgmt rx and tx for bridge peer
Do not expect to receive management frames for the bridge peer
from the firmware. The firmware sends packets to the Host on the
link it received them.

Therefore, add the necessary handling in mgmt_rx_event to drop
the skb buffer if it is received on the bridge peer link.

The same applies to mgmt_tx, where it is not expected for the
Host to send any management frames to the firmware. Hence,
drop the frame.

In ath12k_dp_rx_deliver_msdu(), it is observed that arsta is NULL
and is dereferenced in the bridge peer debug print.
Replace it with peer->link_id.

Fixes: Idae80e0d112 (wifi: ath12k: Avoid acquiring lock on dp rx path)

Change-Id: I22d7e4abbe50f75ac5ff6443b55b54c75f781e51
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>