wifi: ath12k: Remapping Tx-completion-ring4 and rx-error-ring.
Currently,wbm2sw3, wbm2sw4, wbm2sw5 have fixed interrupt number and
fixed interrupt names are assigned to each interrupt in the dts file.

In our current configuration wbm2sw3 mapped to rx-error but it is
assigned to wbm2host-tx-completion-ring4 in the dts.

Due to this rx-error packets are received under wbm2host-tx-completion
-ring4 and tx-completion-ring4 interrupt are not received.

Fix by remapping Tx-completion-ring4 to wbm2sw3 and rx-error-ring to
wbm2sw5 with respect to the dts file and interrupt number.

Patch-dependency: 710-004-ath12k-Add-and-initialize-hw-params-for
-IPQ5332.patch

Change-Id: I5468dacafead5503e7c52ddae9b1f13c57037bae
Signed-off-by: Girish Kumar M P <quic_gmp@quicinc.com>