QSDK: wifi: ath12k: Send low ack event with master netdev index
NL80211_CMD_NOTIFY_CQM command to userspace during SSR recovery.
The ifindex attribute sent along with the NL command is not
recognized by hostapd and is ignored.

This occurs because hostapd is unaware of the AP_VLAN interface created.
Since this event is reported for a connected WDS STA and its sdata is
mapped to the AP_VLAN sdata

To fix this, send the NL command with the Master AP’s ifindex in the
case of WDS.

Patch-dependency: none

Change-Id: I8a254ef9b5a75105b1fd2c081d079afafa2f73bc
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>