ap_sta: parse AP status and link ID properly
With the recent hostap upgrade, the AP will now create a MLD-level socket
under the control interface directory (wlanX). Consequently, the first
entry will no longer be the link-level socket interface. However, the
current logic assumes it is, leading to a failure in reading the link ID
and preventing the repeater AP from being re-enabled.

Fix this now.

Additionally, there is no need to check the EHT flag and link ID, as the
link ID will only exist if it is EHT. Therefore, remove the redundant
EHT checking.

Change-Id: I8789b048e03942cb3430d10a89a10f5fd8b3efa7
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>