wifi: ath12k: fix wrong usage of word mask subscription
The current word mask subscription is only applicable to the RX ring and
not to the monitor destination ring. Using the RX ring MSDU_END structure
in the monitor ring causes invalid data fetching. Fix this issue by
defining the non-compact TLV MSDU_END structure, which is common across all
hardware, and use it in the monitor path. Additionally, remove the
sw_frame_group_id member from the MSDU_END, and correct the MSDU_END word mask
value.

Patch-dependency: 881-wifi-ath12k-add-ethtool-support-for-rx-and-tx-stats.patch

Fixes: (Ie4f1db52c92) wifi: mac80211/ath12k: add support to ethtool stats

Change-Id: Ide93dd7bdb92ea7978aedc431ef20059295420a8
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>