QSDK: wifi: ath12k: Update encap/decap to ethernet for STA
In monitor + station scenario, DS iperf traffic doesn't go through.
This is due to encap type for both monitor and STA is set to
native wifi.
In DS, we update the vif offload type from mac80211 for AP
mode only. Add support to update the offload type and hence
encap to ethernet for STA as well.

Patch-dependency: 951-c-QSDK-wifi-ath12k-support-DS-for-SLO-and-MLO-AP-modes.patch

Change-Id: I2a4f5ce6fb159f0d94b6703e9c43bddde753ed0b
Signed-off-by: Neha Bisht <quic_nbisht@quicinc.com>