wifi: ath12k: Optimize bridge vdev creation logic
Currently bridge vdevs are created immediately after the creation of
the second link in an MLD. However bridge vdevs are not needed in
cases where only the adjacent devices are present in an MLD
combination that is getting configured.

To address this behavior, optimize the code to create the bridge vdevs
only when the links in an MLD combination getting configured, needs a
bridge vdev. This optimization reduces memory consumption of bridge
vdevs and decreases the execution time of vdev creation and deletion
during wifi up/down and ML reconfiguration.

Patch-dependency: 961-01-wifi-ath12k-n-link-mlo-bridge-vdev-basic-handling.patch

Change-Id: Ibc3c6610dbe0a224d4a7a7874a0ea3654b7025a3
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>