wifi: ath12k: Fix ppdu memset issue to avoid stale values
Currently, the mpdu_retry and errmap members are located after the
cmn_mpdu_info member in the struct hal_rx_mon_ppdu_info. When clearing the
existing ppdu_info to avoid a bulk length memset, the size is calculated
by subtracting the size of the ppdu_info struct from the size of the
cmn_mpdu_info struct. Because mpdu_retry and errmap are at the end, they
contain junk values, and the first 64 bits of cmn_mpdu_info are
unintentionally cleared. To fix this, move mpdu_retry and errmap
before the cmn_mpdu_info member.

Fixes : (I1c4bf9c0f36) wifi: cfg80211/ath12k: add support to count rx retries
Fixes : (Ie4f1db52c9) wifi: mac80211/ath12k: add support to ethtool stats

Patch-dependency: 869-b-wifi-ath12k-add-support-to-count-rx-retries.patch

Change-Id: I63552314db7753dea5ff25e3274ce7a04bc1dd90
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>