wifi: ath12k: Increment num_rules by 2 for weather radar reg_rules
When a 5 GHz AP is configured for the country Botswana
(country code BW), the channels 5745 to 5765 are disabled.

In ath12k_reg_build_regd, if the current regdomain is ETSI and if the
current reg rule in iteration contains the 5 GHz weather radar
channels, num_rules is incremented by 1, and an updated regd is
reallocated with num_rules. Since 1 reg_rule received from the target
is split into 3 reg_rules, ideally, num_rules should be incremented
by 2. However, since it is not incremented by 2, the updated regd is
reallocated with memory for 1 less reg_rule, resulting in the last
reg_rules that enable support for channels 5745 to 5765 not being filled.

To address this issue, increment num_rules by 2.

Tag: REGULATORY
Fixes: I0fe49b5f353 ("wifi: ath12k: Increment num_rules by 2 only if weather channels are present")

Change-Id: I690c79dd7736895c84a3021b364b89bfb96e4d65
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>