QSDK: wifi: mac80211: Fix puncturing bitmap during channel switch
Currently, when we try to set puncturing bitmap using hostapd_cli command.
Driver gets the proper values for all parameters, but in function
__ieee80211_channel_switch we have below condition
if (params->punct_bitmap && !sdata->vif.bss_conf.eht_support)
Here, bss_conf.eht_support shows 0 however eht is supported. Due to this
we are not proceeding for puncturing assuming eht is not supported. In beacon
we can observe that bitmap is not present.

Ext Tag: EHT Operation (802.11be D3.0)
     Ext Tag length: 5 (Tag len: 6)
     Ext Tag Number: EHT Operation (802.11be D3.0) (106)
     EHT Operation Parameters: 0x00
          .... ...0 = EHT Operation Information Present: False
          .... ..0. = Disabled Subchannel Bitmap Present: False
          .... .0.. = EHT Default PE Duration: False
          .... 0... = Group Addressed BU Indication Limit: False
          ..00 .... = Group Addressed BU Indication Exponent: 0
          00.. .... = Reserved: 0x0
          Basic EHT-MCS And Nss Set: 0x00000011

We are currently using the vif's bss_conf which is for non-MLO
use only (deflink). In order to work with MLO, appropriate link
conf should be used.

Patch-dependency: none

Fixes: Ia4aac0ca9b3 ("wifi: ath12k: Upgrade ath12k from backport-6.5-rc3 to backport-6.6")

Change-Id: I718341c07427c7364a6ae281f3a4e71b60517034
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>