UPSTREAM:PLATFORM: FR101432: Upstream userpd model for AHB device
With WCSS Secure PIL RPROC driver all the userPD needs to be loaded
and booted by ath12k driver. RootPD is completely controlled by
RPROC driver. Earlier IPQ5332 and IPQ5424 was considered to be rootPD
and only QCN6432 was considered to be userPD. With this new implementation
Q6 is considered to be rootPD and all radios attached to it are
considered to userPDs.
With this new design userPD is loaded, authenticated and power up by
ath12k driver. Similarly, powered down by ath12k driver.
Rproc notification for rootPD is moved from core to AHB specific as
it is apllicable only for AHB. UserPD notifiers are completely as
it is completely handled by Ath12k driver.
In case of UserPD crash, driver receives fatal interrupt, restart of
that userPD is queued in the interrupt handler. Incase of rootPD
crash, fatal rootPD notifier gets invoked from where restart is queued
for all radios, but all are not powered on. All userPDs are powered
on after rootPD is recovered, from rootPD state notifier.
Add support to collect userPD dumps for mode0 and mode1 in ELF format.
Remove userpd_id in ab as this is AHB specific and move it to ab_ahb.

Change-Id: Iec9a78a4477bbacad1e5600059680712d027742c
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>