wifi: ath12k: Add number of rx chains in peer nss calculation
rx chain is not considered in peer nss calculation and the nss
calculation is based on mcs rate and rx nss of sta link.
Due to this, higher value of rx chain is set and it leads to
performance discrimination issue. Add number of rx chain in
peer nss calculation to resolve this issue.

patch-dependency: none

patch-work: none

Change-Id: I16da082aa8b2034545da45b1799afb07706814df
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>