QSDK: ath12k: Remove redundant ppe_vp_type check
In vendor_wifi_config_handler, in case a new updated
ppe_vp_type is received from hostapd then we need to
update the ppe_vp_type config. This needs to be done
for scenarios where vp_config is coming late from
vendor command and vp alloc happens before that only.
But in the current code, we skip updating the ppe_vp_type
due to first setting wdev->ppe_vp_type to new ppe_vp_type
and returning later if it is equal
Remove ppe_vp_type change check in vendor command
handler as this is redundant, as the ppe_vp_type in
wdev was already updated just before this check.

fixes: I66a4d6595 (QSDK: wifi: ath12k: Set wdev VP type at start in vendor command)

Patch-depdendency: 971-QSDK-wifi-ath12k-Update-ppe-vp-config-with-new-ppe_v.patch

Change-Id: Ie83fb0361b4931692965aaa139d5d9d1f6bcca91
Signed-off-by: Pavankumar Nandeshwar <quic_pnandesh@quicinc.com>