wifi: mac80211: Remove warn on from ieee80211_recalc_chanctx_chantype
Currently, in this subroutine compat variable is assiged to NULL by
default. The code checks for compatability on cfg80211 chandef for all
interfaces exist at the time of assignment of link chanctx.

The issue arises when not all interfaces are running or are
of type AP_VLAN, triggering the WARN_ON_ONCE(!compat).

Fix by removing the WARN_ON_ONCE for !compact check under these
conditions, as it is more likely to be encountered.

Fixes: I7d835a0278a ("wifi: ath12k/mac80211: Allow removal of interfaces
during recovery")

Change-Id: I439c14d59196f608f43e6f7958780c35692c36de
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>