UPSTREAM:DP:wifi:ath12k: Fix KW issue ret used unitialized in ath12k_dp_srng_ppeds_setup()
Currently in the ath12k_dp_srng_ppeds_setup() function, during the
if check for dp->ppeds_comp_ring.tx_status, the ret value is not
assigned to anything.
If this condition is true, the ret value is returned without being
allocated any value.

Hence to fix this issue, ret is assigned ENODATA during the if check.

Fixes: 3a0fa36006b3 (wifi: ath12k: Enable UMAC reset support for DS mode)

Change-Id: I3acba843e1d31266da022b5923598c999a26b926
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>