UPSTREAM:PLATFORM:KW:wifi:ath12k: Fix out-of-boundary access in pdev index assignment
The current implementation uses num_hw with a maximum size of
ATH12K_GROUP_MAX_RADIO(14) to determine the pdev index.

However, the actual size of pdevs is based on the ab->num_radios
which has an array index limit of 4.This discrepancy can lead to
out-of-boundary access when the pdev index exceeds the valid range.

Fix this by resetting the pdev index count when it exceeds the
number of radios for the current chip

Fixes 0ba223e170f5e(wifi: ath12k: fix mapping phy names and wlan instance with bands)
Patch-dependency: 856-wifi-ath12k-fix-mapping-phy-names-with-bands.patch
Patch-work:       [None]

Change-Id: I3f2caa847ece8d6b9d4bfcdefef41b72cf94a77a
Signed-off-by: Aishwarya R <quic_aisr@quicinc.com>
Signed-off-by: Arulanbu Balusamy <quic_abalusam@quicinc.com>