UPSTREAM:DP:wifi:mac80211: Fix KW issue for link_id might be used uninitialized
Currently, in the ieee80211_rx_handle_packet() function, the link_id is
assigned if valid_links and wiphy->num_hw are true.
However, when wiphy->num_hw is false, the link_id is assigned to
the previous link_id.
This could result in link_id being used uninitialized.

To prevent this, link_id is now initialized to -1 by default.

Fixes: db8a7fe2ddef (784-mac80211-fix-skb-and-sdata-mismatch.patch)

Change-Id: I1154a91c180a21062929122f197a93e48157aa67
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>