UPSTREAM: CP: wifi: ath12k: Fix signal strength in station dump
WMI_TLV_SERVICE_HW_DB2DBM_CONVERSION_SUPPORT is the service bit that
supports/unsupports conversion from dB to dBm. dB values are usually
positive and dBm values are negative because dBm is the power relative
to 1mW and the power may vary.

The RSSI that is received from the firmware is positive because the
firmware just gives the absolute signal strength. Host checks if the
service bit is supported and if it is, it adds the received values from
firmware with RSSI offset and passes the resulting signal strength in
dBm to application layer. If not, it relays the values given by
firmware, without making changes, to the application layer.

The service bit WMI_TLV_SERVICE_HW_DB2DBM_CONVERSION_SUPPORT is not
supported in QCN firmware, but is supported in WCN firmware. So, the
RSSI from the firmware is relayed to application layer without adding
noise floor which results in positive signal value.

Add RSSI offset received from firmware to the received signal strength from firmware if the service bit is not supported.

Fixes: f9d96f933c6b [902-wifi-ath12k-Fetch-per-chain-rssi-from-firmware.patch]
Patch-dependency: 902-wifi-ath12k-Fetch-per-chain-rssi-from-firmware.patch
Patch-work: [None]

Change-Id: I3d9fd9cf1e724b22e607e461c5d995dd931c3274
Signed-off-by: Roopni Devanathan <quic_rdevanat@quicinc.com>