UPSTREAM:CP: wifi: ath12k: Reset primary link ID after FW recovery
Initially, the AP in WDS STA mode connects to the WDS AP using two links (MLO).
Following this, a firmware (FW) crash is initiated as part of the test case,
causing the FW to crash and then recover with all radios.
After the FW Mode 0 recovery, the host driver brings up the links
one by one to re-associate with the WDS AP.

During this process, the WDS STA attempts to bring up the first link
and associate with the WDS AP. At this point, the primary link ID
retains its old value, which needs to be reset after the FW recovery process.
Consequently, the old primary link ID cannot set the primary UMAC ID,
which must be sent to the FW during peer association.
To resolve this, reset the primary link ID and associated link ID
to the new link ID during reconfiguration.

Patch-dependency: 801-04-wifi-ath12k-Add-support-for-Mode-1-recovery-cache.patch
Patch-work: None

Change-Id: I1891d3672624253ce357697ac1769386aaa578b0
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>