UPSTREAM: CP: wifi: ath12k: Update wsi load info stats for link removal
when a link is getting removed in AP, the corresponding station(arsta)
link is removed from AP. wsi load info stats needs to be updated
because there is change in number of links in ahsta and there is
change in wsi load. Add changes for updating wsi load info stats.

Change-Id: I89b9e308e78e50642d7fbba9a73700671875df4e
Signed-off-by: Mahendran P <quic_mahep@quicinc.com>