UPSTREAM:DP: wifi: ath12k: Assign rcu pointer after completing the peer
After a firmware reset, the AP attempts to recover the firmware one by one in recovery. During this recovery process, the AP flushes the peer list and reassigns peers to the list. However, during this reassignment,
it fails to update the peers with the correct STA values,
leading to a kernel crash. To prevent this, peers are now assigned
to the peer list only after they are fully initialized with the correct values.

Patch-dependency: None
Patch-work: None

Change-Id: If703d833762b4b692f60e8d079bf15c1ffd0845a
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>