wifi: ath12k: drop corrupted packets for invalid band from WBM error ring
Currently, warn_on is getting trigger due to the presence of
corrupted packets associated with an invalid band from the
WBM error ring.
This issue can lead to unnecessary warnings and potential disruptions.

Hence, add a check to drop the packets if it has invalid band and
originating from the WBM error path.
This will ensure corrupted packets are dropped and prevent the warn_on
trigger.

patch-dependency: none

Change-Id: Ic5590c618b07296dafdfa5b2676fa0a0291390d5
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>