UPSTREAM:CP:mac80211.sh: Avoid the race condition during hostapd initialization
For MLO hostapd is triggered from the netifd, after all the cfg files
are created in /var/run/ according to wireless config file. In
some cases, hostapd was called twice due to asynchronous calls from
netifd, causing the hostapd_cfg file to be created for all radios on the
second attempt. This led to hostapd instances being invoked twice.

So to avoid this race condition introducing the flock for the hostapd
invoking part.

Change-Id: I85c08477e6fcf4ff9f16a07a504acb6160569cf8
Signed-off-by: Arulanbu Balusamy <quic_abalusam@quicinc.com>