QSDK:CP:WAR for TCL HUNG Firmware issue
During PDT stress test, while executing wifi restart firmware is not
able to send peer unmap/peer delete response for the peer delete sent
from the host. This is due to TCL hung where in firmware is not able to
respond to peer delete.

Firmware suggested workaround is to send a boolean flag to indicate
whether the peer delete is triggered due to wifi down for taking
respective actions in their side. Below is the changes done for
informing the firmware about the same.

        * Reused existing wdev flags for gettting interface down state from mac80211.
In driver, from arvif, wdev is accessed hence passing arvif to peer delete routine.
        * Last byte of memory allocated for peer mac address is used for sending
the information to the firmware.

Patch-dependency: 691-0001-cfg80211-mac80211-add-support-for-channel-switch-wit.patch
Patch-work: none

Change-Id: If4b86f28dc5176cfa3bc0b3317f731580d669e33
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>