QSDK:CP: wifi: ath12k: Fix SDWF UL traffic and sleep while atomic issue
In SDWF UL traffic scenarios, during traffic start and stop cases,
ath12k UL callback is called with certain parameters, with which
the UL parameters is derived in ath12k and sent to firmware. Address
the issue in the "SUPPORT_ECM_REGISTRATION" availability sanity check
in hw->flags, where the UL callback should be allowed when
"SUPPORT_ECM_REGISTRATION" is registered from ath12k driver. However
the issue is, it is not getting allowed when the flag is registered.
Also with the change in design to call the ath12k callbacks with
"ieee80211_vif" instead of "netdev", two new sanity checks are needed.
Add the sanity checks to prevent NULL pointer crash.

Bug on sleep while atomic issue is observed when sawftx stats are
requested from userspace. This issue is due to a rcu_lock implementation
in a mutex_lock code area. Address this issue by completing the RCU
lock session since it is no longer needed.

Fixes:
	I6c4507857f0 (QSDK: wifi: Prevent ath12k from using net devices not belonging to the driver)
	I337f28e36b2 (wifi: ath12k: convert spinlock `base_lock` to RCU lock for improved concurrency)

Change-Id: Ibd14074d34c0f55f7335ac0a4f5851ce05108f18
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>