QSDK: wifi: ath12k: Fix KW issue vlan_iface dereferenced without checking for NULL
In the ath12k_mac_op_add_interface() function, the vlan_iface pointer
is dereferenced without verifying if it is NULL, even though it can be
NULL as indicated by a previous condition check.

Hence to fix this, check vlan_iface is not NULL before proceeding with
the next condition.

fixes :I014d56299d7 (QSDK: wifi: ath12k: Enable DS support for WDS AP_VLAN interface)

Change-Id: Icbb43d4f0b80fc441b49a9dd85d026041d87ab05
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>