QSDK:PROTOCOL:Add support for setting 00(Global mode) country code
1) When country code is set to 00, only 2GHz rules are received from the
FW. 5GHz and 6GHz rules are null for country 00. When ath12k_regd_update
is called for country 00, since the API updates the regd only using 6GHz
rules, regd update never happens.

To address this, add an exception for country code 00 to bypass waiting
for 6GHz pdev and proceed with the regd_update with 2G pdev.

2) After successfully setting country code to 00, an attempt to switch to
a country that is the default_regd fails. This is because there is a check in
ath12k_reg_handle_chan_list where new_regd is checked for NULL and the
default_regd is compared with the new country. In case of country code
00, the new_regd is not updated as regd update fails for 5GHz and 6GHz
pdevs. Therefore, new_regd remains NULL and ath12k_reg_handle_chan_list
returns a failure when the new country being set is same as the
default_regd.

To address this, add an exception for country code 00 to bypass this
check and proceed with the regd update.

Change-Id: I8abe898f312826fbee732b49702d0acbc27bf90d
Signed-off-by: Amith A <quic_amitajit@quicinc.com>