wifi: DP: ath12k: Fix freeing skbs in scatter list of wbm err path
In WBM err path the skbs are put into scatter list until continuation
bit is not set. Once the continuation bit not set then then the
skbs in the scatterlist moved to msdu list for processing. In
some condition like when rxdesc is not available then these scatterlist skbs
may not be put it into msdu list for processing and these scatter list
skbs won't be freed up and that is causing memleak in long run cases.

Change-Id: Iaa9d90a7651c98b43f0cf099a3bc00d6dfe64726
Signed-off-by: Tamizh Chelvam Raja <quic_tamizhr@quicinc.com>