wifi:ath11k: Handle duplicate descriptor case in mon dest ring
When the software cookie from the MSDU link descriptor matches the
previous software cookie in the monitor destination ring, the physical
address for the corresponding ring is not stored. This results in invalid
memory access when a duplicate entry is detected in the monitor
destination ring. Fix it by caching the last physical address and
comparing it with the new physical address. Skip the monitor destination
ring processing if they are equal.

Patch-dependency: None

Change-Id: I9cd4f913009be74b5ae5dac8ffc8ee80327c499c
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>