QSDK:DP wifi: ath12k: Fix KW issue for wdev might be NULL
Currently, in the ath12k_mac_op_add_interface() function, wdev is
assigned a return value of function which can be NULL, but the
code does not check for NULL pointer value and proceeds to
dereference it. Fix this by checking wdev for NULL value.

Fixes: 7616fed23b04 (QSDK: wifi: ath12k:  Handle vendor command for
interface offload type)

Patch-depdendency: 951-c-QSDK-wifi-ath12k-support-DS-for-SLO-and
 MLO-AP-modes.patc

Change-Id: Id94d5e170e2f0fbd2320cdb6ba50ba1969acfb2c
Signed-off-by: Aditya Rajendra Gawali<quic_agawali@quicinc.com>