UPSTREAM:CP: wifi: ath12k: Add support for bridge VAP in STA mode
In STA mode, similar to AP mode, a bridge vdev is added when links
are established between diagonally opposite devices in an MLD.
However, in STA mode, a single bridge vdev suffices for MLO
communications between these diagonally opposite devices. Therefore,
only one bridge vdev will be created in the adjacent device of the
diagonally opposite devices.

Here the adjacent devices where the bridge is created are
determined using the “device_bitmap” stored in “ahvif”. This bitmap
represents the device_idx of the vdevs created under that “ahvif”.
With this information, the devices where a bridge vdev can be
created are selected based on the adjacent device information of the
currently created vdevs. If the first choice of the adjacent
device cannot create a bridge vdev due to max bridge vdevs allocated
or any other errors, it will move to the next adjacent device for
bridge vdev creation. And even the second adjacent device cannot
create bridge vdev, it will exit gracefully as in AP mode.

This patch brings in these Bridge VAP handlings in STA mode.

Patch-dependency:
        961-02-wifi-ath12k-n-link-mlo-bridge-vdev-create-start.patch

Tag: N_LINK_MLO

Change-Id: I696c37887cf13a3859b4075fcab30c85905e10a7
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>