UPSTREAM:PROTOCOL wifi: ath12k: Fix get_tx_link
Currently, when obtaining the tx link_id in the case of MLO,
if the STA is not available, the link ID passed or the default vif
link is used.

In the case of ML Reconfiguration Link Removal, if link 0 or
the default link is removed, this leads to a multicast traffic drop.

Fix this issue by taking first available link instead of default vif
link.

Change-Id: Iaffce2837ba36ec5a1a35822b328324b12918aad
Signed-off-by: Yuvarani V <quic_yuvarani@quicinc.com>