UPSTREAM:CP: wifi: ath12k: Support SSR Mode1 for bridge VAP in STA mode
Add support to handle bridge VAP in STA mode when SSR mode 1 is
triggered. When SSR mode 1 is triggered in a device where bridge VAP is present, during recovery peer assoc and VAP up needs to be sent for that bridge VAP in STA mode. Handle sending peer assoc and VAP up in STA mode during this Mode 1 recovery phase.

Change-Id: I8534c46d34ee15e6b11f8e640a402c02e60c695a
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>