QSDK:CP:wpa_supplicant.uc: Handle radio id and ML station support
Handle radio id and MLD configurations gracefully in ucode
for both wpa_supplicant and hostapd. This will be useful for
remove interface cases

For ML station config, update the station radio details in
wireless config stored in and later use it for frequency calculation.

For ML station, start only one interface else defer the start_pending and
if radio bitmap is present notify to hostapd accordingly.

Co-developed-by: Nagarajan Maran <quic_nmaran@quicinc.com>
Change-Id: Ib4bbcb16c0111ebb3b3086f7af3618d005a0f6e5
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>