QSDK:DP:wifi:ath12k:fix invalid address passing to prefetch in rx path
For prefetch desc_va is directly getting passed without type casting it
to the pointer whereas prefetch expects a pointer type, not an u64
integer. Fix it by type cast the desc_va_next as a pointer to void,
which corrects the type for the prefetch function to ensure prefetch the
data at the stored address in desc_va_next.

Currently in rx path, driver attempts to assign a pointer to a specific
location in a ring buffer. However, it does not correctly handle pointer
arithmetic. The srng->ring_base_vaddr is likely a pointer, and adding
tp directly to it without proper casting can lead to incorrect address
calculation. Fix it by ensuring that the pointer arithmetic is done
correctly by casting srng->ring_base_vaddr to a char pointer first,
then adding the offset tp, and finally casting the result back to
the appropriate pointer type.

Patch-dependency: None

Change-Id: Iac0a1276fd181853dafed19f7df0f86e0d9ec73c
Signed-off-by: Kaviarasu Sanjeevi <quic_kaviaras@quicinc.com>
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>